requirements.txt
paramiko>=2.7.2
cryptography==42.0.5
bcrypt==4.1.2
PyYAML>=5.4.1
pyserial>=3.5
colorama>=0.4.4  # For Windows console color support 